Move debug_print out of unix
