+++ exception-handler/src/unix.rs
@@ -1,22 +1,6 @@
-#[cfg(feature = "debug-print")]
-#[macro_export]
-macro_rules! debug_print {
-    ($s:literal) => {
-        let cstr = concat!($s, "\n");
-        $crate::write_stderr(cstr);
-    };
-}
+mod pthread_intercept;
 
-#[cfg(not(feature = "debug-print"))]
-macro_rules! debug_print {
-    ($s:literal) => {};
-}
-
-/// Writes the specified string directly to stderr. This is safe to be called
-/// from within a compromised context.
-#[inline]
-pub fn write_stderr(s: &'static str) {
-    unsafe {
-        libc::write(2, s.as_ptr().cast(), s.len());
-    }
-}
+// Force this function to be linked, but it shouldn't actually be called by
+// users directly
+#[doc(hidden)]
+pub use pthread_intercept::pthread_create;
